broke samples into separate files
