+++ container-registry/container-analysis/src/main/java/com/example/containeranalysis/PubSub.java
@@ -1,0 +1,124 @@
+/*
+ * Copyright 2018 Google LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.example.containeranalysis;
+
+// [START containeranalysis_pubsub]
+import static java.lang.Thread.sleep;
+
+import com.google.cloud.devtools.containeranalysis.v1beta1.ContainerAnalysisV1Beta1Client;
+import com.google.cloud.devtools.containeranalysis.v1beta1.GrafeasV1Beta1Client;
+import com.google.cloud.devtools.containeranalysis.v1beta1.GrafeasV1Beta1Client.ListNoteOccurrencesPagedResponse;
+import com.google.cloud.devtools.containeranalysis.v1beta1.GrafeasV1Beta1Client.ListOccurrencesPagedResponse;
+import com.google.cloud.pubsub.v1.AckReplyConsumer;
+import com.google.cloud.pubsub.v1.MessageReceiver;
+import com.google.cloud.pubsub.v1.Subscriber;
+import com.google.cloud.pubsub.v1.SubscriptionAdminClient;
+import com.google.containeranalysis.v1beta1.NoteName;
+import com.google.containeranalysis.v1beta1.OccurrenceName;
+import com.google.containeranalysis.v1beta1.ProjectName;
+import com.google.pubsub.v1.ProjectSubscriptionName;
+import com.google.pubsub.v1.ProjectTopicName;
+import com.google.pubsub.v1.PubsubMessage;
+import com.google.pubsub.v1.PushConfig;
+import com.google.pubsub.v1.Subscription;
+import io.grafeas.v1beta1.ListNoteOccurrencesRequest;
+import io.grafeas.v1beta1.ListOccurrencesRequest;
+import io.grafeas.v1beta1.Note;
+import io.grafeas.v1beta1.Occurrence;
+import io.grafeas.v1beta1.Resource;
+import io.grafeas.v1beta1.UpdateNoteRequest;
+import io.grafeas.v1beta1.UpdateOccurrenceRequest;
+import io.grafeas.v1beta1.vulnerability.Details;
+import io.grafeas.v1beta1.vulnerability.Severity;
+import io.grafeas.v1beta1.vulnerability.Vulnerability;
+import io.grpc.StatusRuntimeException;
+import java.io.IOException;
+import java.lang.InterruptedException;
+import java.util.concurrent.TimeUnit;
+
+public class PubSub {
+  /**
+   * Handle incoming Occurrences using a Cloud Pub/Sub subscription
+   * @param subId the user-specified identifier for the Pub/Sub subscription
+   * @param timeout the amount of time to listen for Pub/Sub messages (in seconds)
+   * @param projectId the GCP project the Pub/Sub subscription belongs to
+   * @return number of Occurrence Pub/Sub messages received before exiting
+   * @throws InterruptedException on errors with the subscription client
+   */
+  public static int pubSub(String subId, int timeout, String projectId)
+      throws InterruptedException {
+    Subscriber subscriber = null;
+    MessageReceiverExample receiver = new MessageReceiverExample();
+
+    try {
+      // Subscribe to the requested Pub/Sub channel
+      ProjectSubscriptionName subName = ProjectSubscriptionName.of(projectId, subId);
+      subscriber = Subscriber.newBuilder(subName, receiver).build();
+      subscriber.startAsync().awaitRunning();
+      // Listen to messages for 'timeout' seconds
+      for (int i = 0; i < timeout; i++) {
+        sleep(1000);
+      }
+    } finally {
+      // Stop listening to the channel
+      if (subscriber != null) {
+        subscriber.stopAsync();
+      }
+    }
+    // Print and return the number of Pub/Sub messages received
+    System.out.println(receiver.messageCount);
+    return receiver.messageCount;
+  }
+
+  /**
+   * Custom class to handle incoming Pub/Sub messages
+   * In this case, the class will simply log and count each message as it comes in
+   */
+  static class MessageReceiverExample implements MessageReceiver {
+    public int messageCount = 0;
+
+    @Override
+    public synchronized void receiveMessage(PubsubMessage message, AckReplyConsumer consumer) {
+      // Every time a Pub/Sub message comes in, print it and count it
+      System.out.println("Message " + messageCount + ": " + message.getData().toStringUtf8());
+      messageCount += 1;
+      // Acknowledge the message
+      consumer.ack();
+    }
+  }
+
+  /**
+   * Creates and returns a Pub/Sub subscription object listening to the Occurrence topic
+   * @param subId the identifier you want to associate with the subscription
+   * @param projectId the GCP project to create the subscription under
+   * @throws IOException thrown on errors with the subscription client
+   * @throws StatusRuntimeException if subscription already exists
+   * @throws InterruptedException on errors shutting down the Grafeas client
+   */
+  public static Subscription createOccurrenceSubscription(String subId, String projectId) 
+      throws IOException, StatusRuntimeException, InterruptedException {
+    // This topic id will automatically receive messages when Occurrences are added or modified
+    String topicId = "container-analysis-occurrences-v1beta1";
+    SubscriptionAdminClient client = SubscriptionAdminClient.create();
+    PushConfig config = PushConfig.getDefaultInstance();
+    ProjectTopicName topicName = ProjectTopicName.of(projectId, topicId);
+    ProjectSubscriptionName subName = ProjectSubscriptionName.of(projectId, subId);
+    Subscription sub = client.createSubscription(subName, topicName, config, 0);
+    return sub;
+  } 
+}
+// [END containeranalysis_pubsub]
